fix if plevels are None
